+++ src/bika/lims/content/analysisrequest.py
@@ -1645,27 +1645,44 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
                 return True
         return False
 
+    def getRawReports(self):
+        """Returns UIDs of reports with a reference to this sample
+
+        see: ARReport.ContainedAnalysisRequests field
+
+        :returns: List of report UIDs
+        """
+        return get_backreferences(self, "ARReportAnalysisRequest")
+
+    def getReports(self):
+        """Returns a list of report objects
+
+        :returns: List of report objects
+        """
+        return list(map(api.get_object, self.getRawReports()))
+
     def getPrinted(self):
         """ returns "0", "1" or "2" to indicate Printed state.
             0 -> Never printed.
             1 -> Printed after last publish
             2 -> Printed but republished afterwards.
         """
-        workflow = getToolByName(self, 'portal_workflow')
-        review_state = workflow.getInfoFor(self, 'review_state', '')
-        if review_state not in ['published']:
+        if not self.getDatePublished():
             return "0"
-        report_list = sorted(self.objectValues('ARReport'),
-                             key=lambda report: report.getDatePublished())
-        if not report_list:
+
+        report_uids = self.getRawReports()
+        if not report_uids:
             return "0"
-        last_report = report_list[-1]
+
+        last_report = api.get_object(report_uids[-1])
         if last_report.getDatePrinted():
             return "1"
-        else:
-            for report in report_list:
-                if report.getDatePrinted():
-                    return "2"
+
+        for report_uid in report_uids[:-1]:
+            report = api.get_object(report_uid)
+            if report.getDatePrinted():
+                return "2"
+
         return "0"
 
     @security.protected(View)