+++ src/senaite/core/catalog/report_catalog.py
@@ -37,7 +37,7 @@ INDEXES = BASE_INDEXES + [
 
 COLUMNS = BASE_COLUMNS + [
     # attribute name
-    "getClientTitle"
+    "getClientTitle",
     "getClientURL",
     "getCreatorFullName",
     "getFileSize",